Revert "1st try at making code slimmer"
This reverts commit e4b7e3f6c4c802ab044e99f5c4ae736b35655257.

+++ PMB/firmware/libraries/ADS1115_min/ADS1115_min.cpp
@@ -3,7 +3,7 @@
 #include <ADS1115_min.h>
 
 ADS1115::ADS1115(int i2c_address){
-	address = uint8_t(i2c_address);
+	address = i2c_address;
 }
 
 void ADS1115::init(){
